rename experiment to training run
